+++ src/service-map/index.js
@@ -1,0 +1,10 @@
+export {
+  SERVICE_CATEGORY_NAMES,
+  SERVICE_LIFECYCLE_STAGES,
+  SERVICE_MAP_ERROR_CODES
+} from './constants.js';
+export { ServiceCategory } from './contracts/service-category.js';
+export { ServiceDescriptor } from './contracts/service-descriptor.js';
+export { ServiceLandscapeSnapshot } from './contracts/service-landscape-snapshot.js';
+export { ServiceLandscape } from './landscape/service-landscape.js';
+export { addServiceMap } from './service-registration.js';